feat(cache): reorganize internals
