+++ crates/pollux/src/core/interfaces.rs
@@ -2,7 +2,6 @@
 // SPDX-License-Identifier: MIT
 
 use crate::core::models::{CargoPackage, CrateVeracityLevel};
-use std::path::Path;
 
 pub trait VeracityFactorEvaluation {
     async fn evaluate(&self, cargo_package: &CargoPackage) -> anyhow::Result<bool>;
@@ -13,11 +12,6 @@ pub trait CrateVeracityLevelEvaluation {
 }
 
 pub trait VeracityEvaluationStorage {
-    fn read(&self, cargo_package: &CargoPackage) -> anyhow::Result<CrateVeracityLevel>;
-    fn save(&self, cargo_package: &CargoPackage, veracity_level: CrateVeracityLevel) -> anyhow::Result<()>;
-}
-
-pub trait PackagesResolution {
-    async fn resolve_for_local_project(&self, project_path: &Path) -> anyhow::Result<Vec<CargoPackage>>;
-    async fn resolve_for_crate_package(&self, cargo_package: &CargoPackage) -> anyhow::Result<Vec<CargoPackage>>;
+    fn retrieve_evaluation(&self, cargo_package: &CargoPackage) -> anyhow::Result<CrateVeracityLevel>;
+    fn save_evaluation(&self, cargo_package: &CargoPackage, veracity_level: CrateVeracityLevel) -> anyhow::Result<()>;
 }
