+++ crates/pollux/src/ioc.rs
@@ -6,10 +6,11 @@ use crate::core::evaluators::standalone::{
     BuildReproducibilityEvaluator, CachedExecutionEvaluator, CrateProvenanceEvaluator,
 };
 use crate::infra::caching::CacheManager;
-use crate::infra::caching::filesystem::DirectoryBased;
-use crate::infra::networking::crates::CratesDotIOClient;
-use crate::infra::networking::crates::cargo::{CrateArchiveDownloader, DependenciesResolver};
-use crate::infra::networking::crates::registry::OfficialCratesRegistryEvaluator;
+use crate::infra::caching::evaluations::VeracityEvaluationsCache;
+use crate::infra::networking::crates::OfficialCratesRegistryEvaluator;
+use crate::infra::networking::crates::registry::CratesDotIOClient;
+use crate::infra::networking::crates::resolvers::DependenciesResolver;
+use crate::infra::networking::crates::tarballs::CrateArchiveDownloader;
 use crate::infra::networking::http::HTTP_CLIENT;
 use crate::infra::networking::ossrebuild::OssRebuildEvaluator;
 use crate::infra::networking::{crates, ossrebuild};
@@ -19,13 +20,13 @@ use crate::pollux::actors::PolluxEvaluatorActor;
 pub static MILLIS_TO_WAIT_AFTER_RATE_LIMITED: u64 = 1100;
 
 fn cached_evaluator() -> CachedExecutionEvaluator {
-    let delegate = DirectoryBased::new(CacheManager::get());
+    let delegate = VeracityEvaluationsCache::new(CacheManager::get());
     CachedExecutionEvaluator::FileSystem(delegate)
 }
 
 fn cratesio_client() -> CratesDotIOClient {
     CratesDotIOClient::new(
-        crates::URL_OFFICIAL_CRATES_REGISTRY.to_string(),
+        crates::registry::URL_OFFICIAL_CRATES_REGISTRY.to_string(),
         HTTP_CLIENT.clone(),
         MILLIS_TO_WAIT_AFTER_RATE_LIMITED,
     )
